% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erotica}
\alias{erotica}
\title{AAT examining approach bias for erotic stimuli}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://osf.io/6h2rj/}{osf.io repository}
}
\usage{
erotica
}
\description{
AAT
}
\references{
Kahveci, S., Van Bockstaele, B.D., & Wiers, R.W. (in preparation).
Pulling for Pleasure? Erotic Approach-Bias Associated With Porn Use, Not Problems. DOI:10.17605/OSF.IO/6H2RJ
}
\keyword{datasets}
