% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortools.R
\name{partial.cor}
\alias{partial.cor}
\title{Partial correlation
Compute the correlation between x and y while controlling for z.}
\usage{
partial.cor(x, y, z, use = c("complete.obs", "everything"))
}
\arguments{
\item{x, y, z}{x and y will be correlated while controlling for z}

\item{use}{optional character indicating how to handle missing values (see \link{cor})}
}
\description{
Partial correlation
Compute the correlation between x and y while controlling for z.
}
\examples{
partial.cor(mtcars$mpg,mtcars$cyl,mtcars$disp)
}
